requirements.txt
pandas==2.0.3
requests==2.31.0
pyyaml==6.0.1
awscli==1.32.0
s3fs==2023.12.0
fsspec==2023.12.0
fastparquet==2023.10.1
pyarrow==10.0.1
clickhouse-driver==0.2.7
clickhouse-connect==0.7.0
